requirements.txt
anthropic>=0.40.0
yfinance>=0.2.50
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
markdown>=3.5.0
